requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
prophet>=1.0.0
tensorflow>=2.5.0
plotly>=5.0.0
dash>=2.0.0
pyyaml>=5.4.0
joblib>=1.0.0
openpyxl>=3.0.0
matplotlib>=3.4.0
seaborn>=0.11.0